requirements.txt
asgiref>=3.5.0
Django==4.0.5
django-localflavor>=3.1
Pillow>=9.1.0
pip==22.1.2
psycopg2-binary>=2.9.3
setuptools==62.4.0
six>=1.16.0
sqlparse>=0.4.2
tzdata>=2022.1
wheel>=0.37.1